Avance, parte obligatoria, solo falta corregir el %p

+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: sergio-alejandro <sergio-alejandro@stud    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/10 11:08:23 by sergio-alej       #+#    #+#             */
-/*   Updated: 2025/11/14 07:26:35 by sergio-alej      ###   ########.fr       */
+/*   Updated: 2025/11/16 11:38:31 by sergio-alej      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libftprintf.h"
+#include "ft_printf.h"
 
-size_t	ft_strlen(const char *s)
+int	ft_strlen(const char *s)
 {
-	size_t i;
+	int i;
 
 	if (!s)
 		return (-1);
